feat: add draggable grid directive for handling drag and drop functionality
