+++ src/app/core/admin/admin-entity-list/admin-entity-list.component.ts
@@ -1,5 +1,4 @@
 import {
-  AfterViewInit,
   Component,
   Input,
   OnChanges,
@@ -22,10 +21,8 @@ import { EntityFieldLabelComponent } from "../../common-components/entity-field-
 import {
   CdkDrag,
   CdkDragDrop,
-  CdkDragEnter,
   CdkDropList,
   CdkDropListGroup,
-  DragRef,
   moveItemInArray,
 } from "@angular/cdk/drag-drop";
 import { FaIconComponent } from "@fortawesome/angular-fontawesome";
@@ -35,6 +32,7 @@ import { MatSelect } from "@angular/material/select";
 import { AdminTabsComponent } from "../building-blocks/admin-tabs/admin-tabs.component";
 import { AdminTabTemplateDirective } from "../building-blocks/admin-tabs/admin-tab-template.directive";
 import { ViewTitleComponent } from "../../common-components/view-title/view-title.component";
+import { DraggableGridDirective } from "app/core/common-components/draggable-grids/draggable-grid.directive";
 
 @Component({
   selector: "app-admin-entity-list",
@@ -59,6 +57,7 @@ import { ViewTitleComponent } from "../../common-components/view-title/view-titl
     AdminTabTemplateDirective,
     ViewTitleComponent,
     CdkDropListGroup,
+    DraggableGridDirective
   ],
   templateUrl: "./admin-entity-list.component.html",
   styleUrls: [
@@ -66,14 +65,9 @@ import { ViewTitleComponent } from "../../common-components/view-title/view-titl
     "../admin-entity/admin-entity-styles.scss",
   ],
 })
-export class AdminEntityListComponent implements OnChanges, AfterViewInit {
+export class AdminEntityListComponent implements OnChanges {
   @ViewChild(CdkDropList) placeholder: CdkDropList;
 
-  private target: CdkDropList = null;
-  private targetIndex: number;
-  private source: CdkDropList = null;
-  private sourceIndex: number;
-  private dragRef: DragRef = null;
   @Input() entityConstructor: EntityConstructor;
   @Input() config: EntityListConfig;
 
@@ -91,11 +85,6 @@ export class AdminEntityListComponent implements OnChanges, AfterViewInit {
 
       this.initAvailableFields();
     }
-  }
-  ngAfterViewInit() {
-    const placeholderElement = this.placeholder.element.nativeElement;
-    placeholderElement.style.display = "none";
-    placeholderElement.parentNode.removeChild(placeholderElement);
   }
   /**
    * Config allows to not have columnGroups and by default then display all `columns`,
@@ -125,69 +114,6 @@ export class AdminEntityListComponent implements OnChanges, AfterViewInit {
     this.filters = (this.config.filters ?? []).map((f) => f.id);
   }
 
-  onDropListDropped() {
-    if (!this.target) {
-      return;
-    }
-    const placeholderElement: HTMLElement =
-      this.placeholder.element.nativeElement;
-    const placeholderParentElement: HTMLElement =
-      placeholderElement.parentElement;
-    placeholderElement.style.display = "none";
-    placeholderParentElement.removeChild(placeholderElement);
-    placeholderParentElement.appendChild(placeholderElement);
-    placeholderParentElement.insertBefore(
-      this.source.element.nativeElement,
-      placeholderParentElement.children[this.sourceIndex],
-    );
-
-    if (this.placeholder._dropListRef.isDragging()) {
-      this.placeholder._dropListRef.exit(this.dragRef);
-    }
-    this.target = null;
-    this.source = null;
-    this.dragRef = null;
-    if (this.sourceIndex !== this.targetIndex) {
-      moveItemInArray(this.config.filters, this.sourceIndex, this.targetIndex);
-    }
-  }
-
-  onDropListEntered({ item, container }: CdkDragEnter) {
-    if (container == this.placeholder) {
-      return;
-    }
-    const placeholderElement: HTMLElement =
-      this.placeholder.element.nativeElement;
-    const sourceElement: HTMLElement = item.dropContainer.element.nativeElement;
-    const dropElement: HTMLElement = container.element.nativeElement;
-    const dragIndex: number = Array.prototype.indexOf.call(
-      dropElement.parentElement.children,
-      this.source ? placeholderElement : sourceElement,
-    );
-    const dropIndex: number = Array.prototype.indexOf.call(
-      dropElement.parentElement.children,
-      dropElement,
-    );
-
-    if (!this.source) {
-      this.sourceIndex = dragIndex;
-      this.source = item.dropContainer;
-      sourceElement.parentElement.removeChild(sourceElement);
-    }
-    this.targetIndex = dropIndex;
-    this.target = container;
-    this.dragRef = item._dragRef;
-    placeholderElement.style.display = "";
-    dropElement.parentElement.insertBefore(
-      placeholderElement,
-      dropIndex > dragIndex ? dropElement.nextSibling : dropElement,
-    );
-    this.placeholder._dropListRef.enter(
-      item._dragRef,
-      item.element.nativeElement.offsetLeft,
-      item.element.nativeElement.offsetTop,
-    );
-  }
 
   updateFilters(filters: string[]) {
     this.filters = filters;
