+++ src/app/core/common-components/draggable-grids/draggable-grid.directive.spec.ts
@@ -1,0 +1,8 @@
+import { DraggableGridDirective } from './draggable-grid.directive';
+
+describe('DraggableGridDirective', () => {
+  it('should create an instance', () => {
+    const directive = new DraggableGridDirective();
+    expect(directive).toBeTruthy();
+  });
+});
